Refactor: Mount Data-Management Handlers as a Child Resource (#794)
Step 3 of the plan in docs/issues/local-api-resource-routing-redesign.md,
plus four extractions taken along the way. api/api_resource.py goes from
2,586 to 1,306 lines.

Registration had no way to express "this handler is not part of the public
API" -- every public method became a route in one flat namespace, and the
only lever for hiding one was a leading underscore, which also lied about a
method's Python visibility.

api/admin_resource.py takes the eleven data-management handlers, the eleven
helpers only they reach, prefer_score_tuner, and the module-level constants
and imports no other part of the file used. APIResource mounts it under a
path prefix using the same build_route_table its own registration uses, so
a child cannot end up registered by different rules than its parent.

The seam was computed from the AST, then asserted in tests: five methods
(_reload_engine, _run_query, _serve_static_file, _set_statement_timeout,
_setup_complete) and four handles (_conn_pool, _cache_generation,
_last_import_time, _setup_complete_cache) cross from child to parent. The
child holds a parent reference for those -- a routing boundary, not a
decoupling (that's what the next two commits are for).

advertise=False is passed once at the mount rather than repeated on each
handler, so forgetting it is a property of the single call site rather than
a hole in one route; build_routes_listing filters on the flag so the 404
listing can't become a directory of what's behind the mount.

Four extractions taken along the way, landed first so the admin move lands
against a smaller file: api/utils/routing.py (registration loop, 404
listing), api/utils/site_name.py (Host-header display names),
api/utils/page_rendering.py (template assembly, minification), and
api/utils/caching.py (the settings-aware cached decorator).
