+++ api/tests/test_import_card_by_name.py
@@ -11,6 +11,7 @@ import orjson
 import pytest
 import requests
 
+from api.admin_resource import AdminResource
 from api.api_resource import APIResource
 
 
@@ -28,29 +29,14 @@ class TestImportCardByName(unittest.TestCase):
     def test_import_card_by_name_validates_input(self) -> None:
         """Test that import_card_by_name validates card_name parameter."""
         with pytest.raises(ValueError) as context:
-            self.api_resource.import_card_by_name(card_name="")
+            self.api_resource.admin.import_card_by_name(card_name="")
 
         assert str(context.value) == "card_name parameter is required"
 
         with pytest.raises(ValueError) as context:
-            self.api_resource.import_card_by_name(card_name=None)
+            self.api_resource.admin.import_card_by_name(card_name=None)
 
         assert str(context.value) == "card_name parameter is required"
-
-    def test_import_card_by_name_function_exists(self) -> None:
-        """Test that import_card_by_name method exists and is callable."""
-        assert hasattr(self.api_resource, "import_card_by_name")
-        assert callable(self.api_resource.import_card_by_name)
-
-    def test_scryfall_search_function_exists(self) -> None:
-        """Test that _scryfall_search method exists."""
-        assert hasattr(self.api_resource, "_scryfall_search")
-        assert callable(self.api_resource._scryfall_search)
-
-    def test_upsert_cards_function_exists(self) -> None:
-        """Test that _upsert_cards method exists."""
-        assert hasattr(self.api_resource, "_upsert_cards")
-        assert callable(self.api_resource._upsert_cards)
 
     @patch("requests.Session.get")
     def test_scryfall_search_returns_empty_for_404(self, mock_get: MagicMock) -> None:
@@ -60,7 +46,7 @@ class TestImportCardByName(unittest.TestCase):
         mock_response.raise_for_status.side_effect = requests.HTTPError("404 Not Found")
         mock_get.return_value = mock_response
 
-        result = self.api_resource._scryfall_search(query="name:'NonexistentCard'")
+        result = self.api_resource.admin._scryfall_search(query="name:'NonexistentCard'")
 
         assert result == []
 
@@ -77,7 +63,7 @@ class TestImportCardByName(unittest.TestCase):
         )
         mock_get.return_value = mock_response
 
-        result = self.api_resource._scryfall_search(query="name:'Lightning Bolt'")
+        result = self.api_resource.admin._scryfall_search(query="name:'Lightning Bolt'")
 
         assert result == [{"name": "Lightning Bolt", "cmc": 1}]
         mock_get.assert_called_once_with(
@@ -95,7 +81,7 @@ class TestImportCardByName(unittest.TestCase):
         mock_get.side_effect = requests.RequestException("Network error")
 
         with pytest.raises(ValueError, match="Failed to fetch data from Scryfall API"):
-            self.api_resource._scryfall_search(query="name:'Lightning Bolt'")
+            self.api_resource.admin._scryfall_search(query="name:'Lightning Bolt'")
 
     @patch.object(APIResource, "_run_query")
     def test_import_card_by_name_returns_already_exists_for_existing_card(self, mock_run_query: MagicMock) -> None:
@@ -103,14 +89,14 @@ class TestImportCardByName(unittest.TestCase):
         # Mock _run_query to return existing card
         mock_run_query.return_value = {"result": [{"card_name": "Lightning Bolt"}]}
 
-        result = self.api_resource.import_card_by_name(card_name="Lightning Bolt")
+        result = self.api_resource.admin.import_card_by_name(card_name="Lightning Bolt")
 
         assert result["status"] == "already_exists"
         assert result["card_name"] == "Lightning Bolt"
         assert "already exists in database" in result["message"]
 
     @patch.object(APIResource, "_run_query")
-    @patch.object(APIResource, "_scryfall_search")
+    @patch.object(AdminResource, "_scryfall_search")
     def test_import_card_by_name_returns_not_found_for_missing_card(
         self,
         mock_search: MagicMock,
@@ -123,14 +109,14 @@ class TestImportCardByName(unittest.TestCase):
         # Mock Scryfall API to return empty list (not found)
         mock_search.return_value = []
 
-        result = self.api_resource.import_card_by_name(card_name="NonexistentCard")
+        result = self.api_resource.admin.import_card_by_name(card_name="NonexistentCard")
 
         assert result["status"] == "not_found"
         assert result["search_query"] == '!"NonexistentCard"'
         assert "No cards found for search query" in result["message"]
 
     @patch.object(APIResource, "_run_query")
-    @patch.object(APIResource, "_scryfall_search")
+    @patch.object(AdminResource, "_scryfall_search")
     def test_import_card_by_name_returns_error_for_scryfall_exceptions(
         self,
         mock_search: MagicMock,
@@ -143,14 +129,14 @@ class TestImportCardByName(unittest.TestCase):
         # Mock Scryfall API to raise exception
         mock_search.side_effect = ValueError("API Error")
 
-        result = self.api_resource.import_card_by_name(card_name="TestCard")
+        result = self.api_resource.admin.import_card_by_name(card_name="TestCard")
 
         assert result["status"] == "error"
         assert result["search_query"] == '!"TestCard"'
         assert "Error fetching cards from Scryfall" in result["message"]
 
     @patch.object(APIResource, "_run_query")
-    @patch.object(APIResource, "_scryfall_search")
+    @patch.object(AdminResource, "_scryfall_search")
     @patch("api.card_processing.preprocess_card")
     def test_import_card_by_name_returns_filtered_out_for_invalid_cards(
         self,
@@ -168,7 +154,7 @@ class TestImportCardByName(unittest.TestCase):
         # Mock preprocessing to return None (filtered out)
         mock_preprocess.return_value = []
 
-        result = self.api_resource.import_card_by_name(card_name="TestCard")
+        result = self.api_resource.admin.import_card_by_name(card_name="TestCard")
         assert result == {
             "status": "no_cards_after_preprocessing",
             "cards_loaded": 0,
